+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update open layers to version 10.4.0. #206
+- Update ol-layerswitcher, ol-popup, ol-grid and ol-side-panel to latest versions.
+
 ## [v2.4.0] - 2025-02-15
 
 ### Added